+++ frontend/src/components/Navigation/index.js
@@ -1,10 +1,10 @@
-// frontend/src/components/Navigation/index.js
 import React from "react";
 import { NavLink } from "react-router-dom";
 import { useSelector } from "react-redux";
 import ProfileButton from "./ProfileButton";
 import OpenModalButton from "../OpenModalButton";
 import LoginFormModal from "../LoginFormModal";
+import SignUpFormModal from "../SignUpFormModal";
 // import "./Navigation.css";
 
 function Navigation({ isLoaded }) {
@@ -24,7 +24,10 @@ function Navigation({ isLoaded }) {
 					buttonText='Log In'
 					modalComponent={<LoginFormModal />}
 				/>
-				<NavLink to='/signup'>Sign Up</NavLink>
+				<OpenModalButton
+					buttonText='Sign Up'
+					modalComponent={<SignUpFormModal />}
+				/>
 			</li>
 		);
 	}
